Change uint to size_t

+++ fenics/2d/ns_cylinder/demo_chorin.py
@@ -7,7 +7,7 @@ from dolfin import *
 
 # Load mesh from file
 mesh = Mesh("cylinder_in_channel.xml")
-sub_domains = MeshFunction("uint", mesh, "subdomains.xml")
+sub_domains = MeshFunction("size_t", mesh, "subdomains.xml")
 
 # Define function spaces (P2-P1)
 V = VectorFunctionSpace(mesh, "CG", 2)
