+++ fenics/2d/ns_cylinder/demo_steady.py
@@ -25,7 +25,7 @@ from dolfin import *
 
 # Load mesh from file
 mesh = Mesh("cylinder_in_channel.xml")
-sub_domains = MeshFunction("uint", mesh, "subdomains.xml")
+sub_domains = MeshFunction("size_t", mesh, "subdomains.xml")
 dss = Measure("ds")[sub_domains]
 
 # Define function spaces (P2-P1)
